updated tests fixtures (user tests updated)

+++ apps/user/views/VUser.py
@@ -102,6 +102,25 @@ class VUserDetail(APIView):
         ...
         """
         user = self.get_object(pk)
-        # value = user.id
+
+        if user.id == request.user.id:
+            return Response(
+                "can't delete himself",
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        if user.is_superuser:
+            return Response(
+                'super users cannot be deleted',
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        if user.is_staff:
+            if not request.user.is_superuser:
+                return Response(
+                    'user cannot delete administrators',
+                    status=status.HTTP_400_BAD_REQUEST,
+                )
+
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
